// src/Types/levelObj.ts
export type levelObj = {
  name: string
  id: string
  description: string
  author: string
  playerID: string
  accountID: string
  difficulty: string
  downloads: number
  likes: number
  disliked: boolean
  length: string
  stars: number
  orbs: number
  diamonds: number
  featured: boolean
  epic: boolean
  gameVersion: string
  editorTime: number
  totalEditorTime: number
  version: number
  copiedID: string
  twoPlayer: boolean
  officialSong: number
  customSong: number
  coins: number
  verifiedCoins: boolean
  starsRequested: number
  ldm: boolean
  objects: number
  large: boolean
  cp: number
  difficultyFace: string
  songName: string
  songAuthor: string
  songSize: string
  songID: number
  songLink: string
}